+++ src/dipdup/index.py
@@ -13,6 +13,7 @@ from typing import Dict
 from typing import Generic
 from typing import Iterable
 from typing import Iterator
+from typing import NoReturn
 from typing import Optional
 from typing import Sequence
 from typing import Set
@@ -465,14 +466,14 @@ class OperationIndex(Index[OperationIndexConfig]):
         operations = operation_subgroup.operations
 
         for handler_config in self._config.handlers:
-            operation_idx = 0
+            subgroup_index = 0
             pattern_idx = 0
             matched_operations: Deque[Optional[OperationData]] = deque()
 
             # TODO: Ensure complex cases work, e.g. when optional argument is followed by required one
             # TODO: Add None to matched_operations where applicable (pattern is optional and operation not found)
-            while operation_idx < len(operations):
-                operation, pattern_config = operations[operation_idx], handler_config.pattern[pattern_idx]
+            while subgroup_index < len(operations):
+                operation, pattern_config = operations[subgroup_index], handler_config.pattern[pattern_idx]
                 operation_matched = await self._match_operation(pattern_config, operation)
 
                 if operation.type == 'origination' and isinstance(
@@ -487,12 +488,12 @@ class OperationIndex(Index[OperationIndexConfig]):
                 if operation_matched:
                     matched_operations.append(operation)
                     pattern_idx += 1
-                    operation_idx += 1
+                    subgroup_index += 1
                 elif pattern_config.optional:
                     matched_operations.append(None)
                     pattern_idx += 1
                 else:
-                    operation_idx += 1
+                    subgroup_index += 1
 
                 if pattern_idx == len(handler_config.pattern):
                     self._logger.info('%s: `%s` handler matched!', operation_subgroup.hash, handler_config.callback)
@@ -533,22 +534,26 @@ class OperationIndex(Index[OperationIndexConfig]):
                 storage_type = pattern_config.storage_type_cls
                 storage = deserialize_storage(operation_data, storage_type)
 
-                transaction_context: Transaction[Any, Any] = Transaction(
+                typed_transaction: Transaction[Any, Any] = Transaction(
                     data=operation_data,
                     parameter=parameter,
                     storage=storage,
                 )
-                args.append(transaction_context)
+                args.append(typed_transaction)
 
             elif isinstance(pattern_config, OperationHandlerOriginationPatternConfig):
+                if not (pattern_config.originated_contract or pattern_config.similar_to):
+                    args.append(operation_data)
+                    continue
+
                 storage_type = pattern_config.storage_type_cls
                 storage = deserialize_storage(operation_data, storage_type)
 
-                origination_context = Origination(
+                typed_origination = Origination(
                     data=operation_data,
                     storage=storage,
                 )
-                args.append(origination_context)
+                args.append(typed_origination)
 
             else:
                 raise NotImplementedError
@@ -640,7 +645,7 @@ class BigMapIndex(Index):
                     stack.enter_context(Metrics.measure_level_realtime_duration())
                 await self._process_level_big_maps(big_maps, message_level)
 
-    async def _create_fetcher(self, first_level: int, last_level: int) -> DataFetcher:
+    async def _create_fetcher(self, first_level: int, last_level: int) -> BigMapFetcher:
         big_map_addresses = self._get_big_map_addresses()
         big_map_paths = self._get_big_map_paths()
 
@@ -847,7 +852,7 @@ class HeadIndex(Index):
         super().__init__(ctx, config, datasource)
         self._queue: Deque[HeadBlockData] = deque()
 
-    async def _create_fetcher(self, first_level: int, last_level: int) -> DataFetcher:
+    async def _create_fetcher(self, first_level: int, last_level: int) -> NoReturn:
         raise NotImplementedError('HeadIndex has no fetcher')
 
     async def _synchronize(self, sync_level: int) -> None:
@@ -904,7 +909,7 @@ class TokenTransferIndex(Index):
         if Metrics.enabled:
             Metrics.set_levels_to_realtime(self._config.name, len(self._queue))
 
-    async def _create_fetcher(self, first_level: int, last_level: int) -> DataFetcher:
+    async def _create_fetcher(self, first_level: int, last_level: int) -> TokenTransferFetcher:
         token_addresses: set[str] = set()
         token_ids: set[int] = set()
         from_addresses: set[str] = set()
@@ -1072,7 +1077,7 @@ class EventIndex(Index):
                     stack.enter_context(Metrics.measure_level_realtime_duration())
                 await self._process_level_events(events, message_level)
 
-    async def _create_fetcher(self, first_level: int, last_level: int) -> DataFetcher:
+    async def _create_fetcher(self, first_level: int, last_level: int) -> EventFetcher:
         event_addresses = self._get_event_addresses()
         event_tags = self._get_event_tags()
         return EventFetcher(