+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+* codegen: Include PEP 561 marker in generated packages.
+* codegen: Untyped originations are now correctly handled.
+* codegen: Fixed `alias` config field having no effect on originations.
+* codegen: Fixed optional arguments in generated callbacks.
+* config: Suggest snake_case for package name.
+* index: Do not deserialize originations matched by the `source` filter.
+* index: Wrap storage deserialization exceptions with `InvalidDataError`.
+
 ## [6.3.0] - 2022-11-15
 
 ### Added